Revert move to utils package
