+++ dces-report-service/src/main/java/uk/gov/justice/laa/crime/dces/report/client/ContributionFilesClient.java
@@ -7,8 +7,8 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.reactive.function.client.WebClient;
 import org.springframework.web.service.annotation.GetExchange;
 import org.springframework.web.service.annotation.HttpExchange;
-import uk.gov.justice.laa.crime.dces.report.utils.maatapi.MaatApiClientFactory;
-import uk.gov.justice.laa.crime.dces.report.utils.maatapi.client.MaatApiClient;
+import uk.gov.justice.laa.crime.dces.report.maatapi.MaatApiClientFactory;
+import uk.gov.justice.laa.crime.dces.report.maatapi.client.MaatApiClient;
 
 import java.time.LocalDate;
 import java.util.List;
