+++ dces-report-service/src/main/java/uk/gov/justice/laa/crime/dces/report/maatapi/MaatApiWebClientFactory.java
@@ -1,4 +1,4 @@
-package uk.gov.justice.laa.crime.dces.report.utils.maatapi;
+package uk.gov.justice.laa.crime.dces.report.maatapi;
 
 //import io.jsonwebtoken.Jwts;
 //import io.jsonwebtoken.SignatureAlgorithm;
@@ -20,8 +20,8 @@ import org.springframework.web.client.HttpServerErrorException;
 import org.springframework.web.reactive.function.client.*;
 import reactor.netty.http.client.HttpClient;
 import reactor.netty.resources.ConnectionProvider;
-import uk.gov.justice.laa.crime.dces.report.utils.maatapi.config.ServicesConfiguration;
-import uk.gov.justice.laa.crime.dces.report.utils.maatapi.exception.MaatApiClientException;
+import uk.gov.justice.laa.crime.dces.report.maatapi.config.ServicesConfiguration;
+import uk.gov.justice.laa.crime.dces.report.maatapi.exception.MaatApiClientException;
 
 import java.time.Duration;
 import java.util.UUID;
