+++ dces-report-service/src/test/java/uk/gov/justice/laa/crime/dces/report/client/ContributionFilesClientTest.java
@@ -10,7 +10,7 @@ import org.mockito.Mockito;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ActiveProfiles;
-import uk.gov.justice.laa.crime.dces.report.utils.maatapi.exception.MaatApiClientException;
+import uk.gov.justice.laa.crime.dces.report.maatapi.exception.MaatApiClientException;
 
 import java.time.LocalDate;
 import java.util.List;
