+++ dces-report-service/src/test/java/uk/gov/justice/laa/crime/dces/report/maatapi/MaatApiClientFactoryTest.java
@@ -6,8 +6,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.web.reactive.function.client.WebClient;
-import uk.gov.justice.laa.crime.dces.report.utils.maatapi.MaatApiClientFactory;
-import uk.gov.justice.laa.crime.dces.report.utils.maatapi.client.MaatApiClient;
+import uk.gov.justice.laa.crime.dces.report.maatapi.client.MaatApiClient;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
